refactor(tests): mirror new src/ subfolder layout under tests/ and update import paths
